refactor Card.jsx and some components
